refactor(firmware): the absolute param bus is core ground, not MQTT's
RFC step 1 of 6, second half (docs/rfc-core-and-variants.md §2.11). No
behaviour change; the other move that has to happen before a feature can
leave.

The four absolute channels are part of the module ABI: a pattern that
declares ABSOLUTE_READY reads them out of its InputFrame, so every build
that loads modules carries them. They lived inside core_mqtt.h because
MQTT happened to be the first thing that drove them — which left the
show player calling PatternflowMqtt::applyRemoteParam() to move a knob,
and the sketch calling PatternflowMqtt::fillAbsolute() to build a frame.
Contract code owned by an optional feature.

The bus moves to core_bus.h. MQTT keeps its ~40 in-file references
through using-declarations and four thin forwarders, so it reads as one
driver among several rather than the owner; the show player and the
sketch now call PatternflowBus:: directly.

Proof: the tree builds with -DPF_MQTT_ENABLED=0 -DPF_MQTT_HTTP_ENABLED=0
— shows can drive the knobs on a build with no broker code in it at all.

Verified on hardware with docs/pfst-v2-vectors/a-ease-ramp.pfs, whose
expected trajectory the spec prints: the bus climbed 26 → 46 → 71 → 92
→ 118 → 155 → 182 through the eased segment, held 1000 after it, and
released every channel on stop. Console pages and frame time unchanged.
