+++ firmware/patternflow/src/core_mqtt.h
@@ -41,6 +41,7 @@
 #include <Arduino.h>
 #include "config.h"
 #include "core_encoders.h"
+#include "core_bus.h"
 
 #if PF_MQTT_ENABLED
 #include <WiFi.h>
@@ -128,13 +129,14 @@ inline IPAddress brokerIp;
 inline bool brokerResolved = false;
 inline uint8_t failures = 0;
 
-// Absolute param bus (held until physical encoder motion releases a channel).
-inline bool paramHeld[4] = {false, false, false, false};
-inline uint16_t paramValue[4] = {500, 500, 500, 500};
-inline uint32_t paramHeldAtMs[4] = {0, 0, 0, 0};
+// The absolute param bus lives in core_bus.h - it is module-ABI ground,
+// not an MQTT feature. MQTT is one of its drivers; these aliases keep the
+// in-file references (publishing, status JSON, inventory) reading the same.
+using PatternflowBus::paramHeld;
+using PatternflowBus::paramValue;
+using PatternflowBus::paramHeldAtMs;
 inline uint32_t lastSnapshotPubMs = 0;
-// Ignore encoder chatter briefly after an absolute set (Director spam / noise).
-constexpr uint32_t ABSOLUTE_RELEASE_GRACE_MS = 250;
+using PatternflowBus::ABSOLUTE_RELEASE_GRACE_MS;
 
 constexpr size_t HOST_BYTES = 64;
 constexpr size_t USER_BYTES = 32;
@@ -571,44 +573,20 @@ inline void applyRemotePattern(const char* name) {
   Serial.printf("[MQTT] unknown pattern '%s'\n", name ? name : "");
 }
 
+// Thin forwarders to core_bus.h, kept so existing callers need no edit.
+// New code should call PatternflowBus:: directly.
 inline void applyRemoteParam(int index, long value) {
-  if (index < 0 || index > 3) return;
-  if (value < 0) value = 0;
-  if (value > 1000) value = 1000;
-  paramHeld[index] = true;
-  paramValue[index] = (uint16_t)value;
-  paramHeldAtMs[index] = millis();
+  PatternflowBus::applyRemoteParam(index, value);
 }
 
 inline void releaseAbsolute(int index) {
-  if (index < 0 || index > 3) return;
-  // Physical release only after grace — avoids encoder noise dropping Director holds.
-  if (paramHeld[index] &&
-      (millis() - paramHeldAtMs[index]) < ABSOLUTE_RELEASE_GRACE_MS) {
-    return;
-  }
-  paramHeld[index] = false;
+  PatternflowBus::releaseAbsolute(index);
 }
 
-inline void clearAbsoluteAll() {
-  for (int i = 0; i < 4; ++i) {
-    paramHeld[i] = false;
-    paramHeldAtMs[i] = 0;
-  }
-}
+inline void clearAbsoluteAll() { PatternflowBus::clearAbsoluteAll(); }
 
-// Copy held absolute values into the frame. Call after physical release.
-// When absolute is active, clear deltas / audio flags on that channel so
-// legacy paths cannot fight the Director set-point.
 inline void fillAbsolute(InputFrame& input) {
-  for (int i = 0; i < 4; ++i) {
-    input.paramAbsoluteActive[i] = paramHeld[i];
-    input.paramAbsolute[i] = paramValue[i];
-    if (paramHeld[i]) {
-      input.knobDeltas[i] = 0;
-      input.knobAudioActive[i] = false;
-    }
-  }
+  PatternflowBus::fillAbsolute(input);
 }
 
 inline void applyRemoteMessage(uint8_t* payload, unsigned int length) {
